implement InstPrinter

+++ testcase/llc/complex.c
@@ -1,6 +1,6 @@
 void test()
 {
-for(int i=0, j = *(int*)0x80;i<j;++i)
+for(long i=0, j = *(long*)0x80;i<j;++i)
 {
 *(long*)0x40 *= i;
 }
